Perfetto plugin - Memory optimizations and table renaming (#48)
Made some memory optimizations to help improve the sluggishness in WPA when viewing large traces. Memory optimizations reduce memory usage by about 50%. UI is more responsive now. Optimizations include

Used string interning on the high-count string fields
Switched the variable length argument fields from List to string[]
Changed bunch of longs to ints. These are mostly ID fields. Confirmed they are int or smaller within Perfetto
Table renaming

Removed redundant "Perfetto" from all tables and column config names
Added column configs that graph by StartTime instead of duration for GenericEvents

Added fix for duplicate GUIDs

+++ PerfettoCds/Pipeline/DataOutput/PerfettoCpuCountersEvent.cs
@@ -8,7 +8,7 @@ namespace PerfettoCds.Pipeline.DataOutput
     public readonly struct PerfettoCpuCountersEvent
     {
         // The specific CPU core
-        public long CpuNum { get; }
+        public int CpuNum { get; }
         public Timestamp StartTimestamp { get; }
         public TimestampDelta Duration { get; }
 
@@ -36,7 +36,7 @@ namespace PerfettoCds.Pipeline.DataOutput
         /// <summary>
         /// For populating the current counter values
         /// </summary>
-        public PerfettoCpuCountersEvent(long cpuNum, Timestamp startTimestamp, TimestampDelta duration,
+        public PerfettoCpuCountersEvent(int cpuNum, Timestamp startTimestamp, TimestampDelta duration,
             double userNs,
             double userNiceNs,
             double systemModeNs,
@@ -70,7 +70,7 @@ namespace PerfettoCds.Pipeline.DataOutput
         /// <summary>
         /// When we have a previous event to compare to, we can calculate the percent change
         /// </summary>
-        public PerfettoCpuCountersEvent(long cpuNum, Timestamp startTimestamp, TimestampDelta duration,
+        public PerfettoCpuCountersEvent(int cpuNum, Timestamp startTimestamp, TimestampDelta duration,
             double userNs,
             double userNiceNs,
             double systemModeNs,
