+++ PerfettoCds/Pipeline/DataOutput/PerfettoCpuFrequencyEvent.cs
@@ -13,14 +13,14 @@ namespace PerfettoCds.Pipeline.DataOutput
         // The current frequency of this CPU
         public double CpuFrequency { get; }
         // The specific CPU core
-        public long CpuNum { get; }
+        public int CpuNum { get; }
         public Timestamp StartTimestamp { get; }
         // Type of CPU frequency event. Whether it's an idle change or frequency change event
         public string Name { get; }
         public TimestampDelta Duration { get; }
         public bool IsIdle { get; }
 
-        public PerfettoCpuFrequencyEvent(double cpuFrequency, long cpuNum, Timestamp startTimestamp, TimestampDelta duration, string name, bool isIdle)
+        public PerfettoCpuFrequencyEvent(double cpuFrequency, int cpuNum, Timestamp startTimestamp, TimestampDelta duration, string name, bool isIdle)
         {
             this.CpuFrequency = cpuFrequency;
             this.CpuNum = cpuNum;
